Fix #3 delete fonctionnality

+++ src/main/java/com/esme/spring/faircorp/hello/BuildingControler.java
@@ -69,14 +69,19 @@ public class BuildingControler {
 
         List<Room> list = roomDaoImpl.FindRoomByBuilding(building.get(0));
 
-        for (int i=0;i<list.size();i++) {
-            List<Light> listl = lightDaoImpl.findOnRoomId(list.get(i));
+        if(!list.size().equals(0)){
+          for (int i=0;i<list.size();i++) {
+              List<Light> listl = lightDaoImpl.findOnRoomId(list.get(i));
 
-            for (int j=0;j<list.size();j++) {
-                lightDao.deleteById(listl.get(j).getId());
-            }
+              if(!listl.size().equals(0)){
+                for (int j=0;j<list.size();j++) {
+                    lightDao.deleteById(listl.get(j).getId());
+                }
+              }
 
-            roomDao.deleteById(list.get(i).getId());
+              roomDao.deleteById(list.get(i).getId());
+          }
         }
+
         buildingDao.deleteById(id);}
 }
